requirements.txt
typesense
psycopg2-binary
python-dotenv